+++ src/handler.c
@@ -737,7 +737,7 @@ int get_curr_stat( CHAR_DATA *ch, int stat )
 
     else
     {
-	max = pc_race_table[ch->race].max_stats[stat] + 4;
+	max = race_table[ch->race].max_stats[stat] + 4;
 
 	if (class_table[ch->class].attr_prime == stat)
 	    max += 2;
@@ -759,7 +759,7 @@ int get_max_train( CHAR_DATA *ch, int stat )
     if (IS_NPC(ch) || ch->level > LEVEL_IMMORTAL)
 	return 25;
 
-    max = pc_race_table[ch->race].max_stats[stat];
+    max = race_table[ch->race].max_stats[stat];
     if (class_table[ch->class].attr_prime == stat)
     {
         if (ch->race == race_lookup("human"))